requirements.txt
Flask==2.3.3
opencv-python
numpy
pytesseract
Pillow
num2words
mysql-connector-python
